chore: refactor AccessibilityOptions
